chore: native wayland support
Electron can now be built with Wayland support.

One problem is that it does not support drawing a title bar for any applications using it.

according to: https://github.com/microsoft/vscode/issues/124202#issuecomment-847827357

> This is a well known issue. Gnome under Wayland does not support server side decorations, which is a problem for all Electron apps that don't implement client side decorations (which VSCode does).

ref: https://github.com/electron/electron/issues/27016

On vscode you can fix it using this setting:

```
"window.titleBarStyle": "custom"
```

so here we following the fix up